+++ src/fractal_task_tools/_create_manifest.py
@@ -66,9 +66,13 @@ def create_manifest(
     TASK_LIST: list[_BaseTask] = getattr(task_list_module, "TASK_LIST")
 
     # Load INPUT_MODELS
-    INPUT_MODELS = getattr(task_list_module, "INPUT_MODELS", None)
-    if INPUT_MODELS is not None:
-        logging.warning("`INPUT_MODELS` is deprecated, and its usage has no effect")
+    try:
+        INPUT_MODELS = getattr(task_list_module, "INPUT_MODELS")
+    except AttributeError:
+        INPUT_MODELS = []
+        logger.warning(
+            "No `INPUT_MODELS` found in task_list module. Setting it to `[]`."
+        )
 
     # Load AUTHORS
     try:
@@ -120,6 +124,7 @@ def create_manifest(
                 schema = create_schema_for_single_task(
                     executable,
                     package=package_name,
+                    pydantic_models=INPUT_MODELS,
                 )
 
                 validate_arguments(