+++ CHANGELOG.md
@@ -1,13 +1,12 @@
 # 0.4.0 (unreleased)
 
-FIXME: Info about class-attribute docstrings
-
 * Schema generation:
     * Compute `default` as specified in `default_factory`, when possible (\#75, \#93).
     * Take into accounts `default` values set through `Field` (\#93).
     * Transform single-element `"allOf": [{"$ref": X}]` arrays into single `"$ref": X` key-value pair (\#79).
     * Support providing `description` through `Field`, with priority over docstrings (\#89).
     * Propagate checks for forbidden non-`null` defaults of nullable properties to nested properties (\#103).
+    * Set `use_attribute_docstrings=True`, so that Pydantic-model-attributes docstrings are used as `description`s with priority over `"Attributes:"` docstrings (\#115).
 * CLI commands:
     * Add `--verbose` option to `fractal-manifest check` (\#80).
     * In `fractal-manifest check`, display all manifest mismatches rather than the first one only (\#88).